use ajax load room

+++ src/main/java/vn/framgia/dao/IBookingDAO.java
@@ -1,10 +1,10 @@
 package vn.framgia.dao;
 
-import java.util.List;
-
 import vn.framgia.model.Booking;
 
-public interface IBookingDAO extends IGenericDAO<Booking, Long>{
+import java.util.List;
+
+public interface IBookingDAO extends IGenericDAO<Booking, Integer>{
 	public List<Booking> findAllOrderClientId();
 	public List<Booking> findBillByNameClient(String nameClient);
 }
